// test/TestingExample.Website.FunctionalTests/packages/scenario-builder/src/client/types.gen.ts
export type CreateDocumentRequestModel = {
    template: null | ReferenceByIdModel;
    documentType: ReferenceByIdModel;
    parent?: null | ReferenceByIdModel;
    id?: null | string;
    values: Array<DocumentValueModel>;
    variants: Array<DocumentVariantRequestModel>;
};
// ...
export type DocumentValueModel = {
    culture?: null | string;
    segment?: null | string;
    alias: string;
    value?: unknown;
};

export type DocumentValueResponseModel = {
    editorAlias: string;
    culture?: null | string;
    segment?: null | string;
    alias: string;
    value?: unknown;
};

export type DocumentVariantItemResponseModel = {
    id: string;
    flags: Array<FlagModel>;
    state: PublishableVariantStateModel;
    name: string;
    culture?: null | string;
};

export type DocumentVariantRequestModel = {
    culture?: null | string;
    segment?: null | string;
    name: string;
};
// ...
export type ReferenceByIdModel = {
    id: string;
};